# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CancelRequest": {
        "properties": {
          "at_period_end": {
            "default": true,
            "title": "At Period End",
            "type": "boolean"
          }
        },
        "title": "CancelRequest",
        "type": "object"
      },
...
      "GaugeIn": {
        "description": "Set a gauge to `value`, or move it by `delta` \u2014 exactly one of the two.",
        "properties": {
          "delta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "metric"
        ],
        "title": "GaugeIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanChangeRequest": {
        "properties": {
          "plan_key": {
            "minLength": 1,
            "title": "Plan Key",
            "type": "string"
          }
        },
        "required": [
          "plan_key"
        ],
        "title": "PlanChangeRequest",
        "type": "object"
      },
      "PlanFeatureRead": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "feature_key": {
            "title": "Feature Key",
            "type": "string"
          }
        },
        "required": [
          "feature_key",
          "enabled"
        ],
        "title": "PlanFeatureRead",
        "type": "object"
      },
      "PlanLimitRead": {
        "properties": {
          "limit_value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Value"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "soft_limit_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soft Limit Ratio"
          }
        },
        "required": [
          "metric",
          "limit_value",
          "soft_limit_ratio"
        ],
        "title": "PlanLimitRead",
        "type": "object"
      },
      "PlanRead": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "features": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlanFeatureRead"
            },
            "title": "Features",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interval": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval"
          },
          "is_custom": {
            "title": "Is Custom",
            "type": "boolean"
          },
          "is_public": {
            "title": "Is Public",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "limits": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlanLimitRead"
            },
            "title": "Limits",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Cents"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "trial_days": {
            "title": "Trial Days",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "description",
          "price_cents",
          "currency",
          "interval",
          "is_public",
          "is_custom",
          "trial_days",
          "sort_order"
        ],
        "title": "PlanRead",
        "type": "object"
      },
...
      "SubscriptionRead": {
        "properties": {
          "cancel_at_period_end": {
            "title": "Cancel At Period End",
            "type": "boolean"
          },
          "canceled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canceled At"
          },
          "current_period_end": {
            "format": "date-time",
            "title": "Current Period End",
            "type": "string"
          },
          "current_period_start": {
            "format": "date-time",
            "title": "Current Period Start",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanRead"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "plan_snapshot": {
            "additionalProperties": true,
            "title": "Plan Snapshot",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trial_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Ends At"
          }
        },
        "required": [
          "id",
          "organization_id",
          "plan_id",
          "plan",
          "status",
          "current_period_start",
          "current_period_end",
          "trial_ends_at",
          "cancel_at_period_end",
          "canceled_at",
          "plan_snapshot"
        ],
        "title": "SubscriptionRead",
        "type": "object"
      },
...
      "TrialStartRequest": {
        "properties": {
          "plan_key": {
            "minLength": 1,
            "title": "Plan Key",
            "type": "string"
          }
        },
        "required": [
          "plan_key"
        ],
        "title": "TrialStartRequest",
        "type": "object"
      },
      "UsageBatchIn": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/UsageEventIn"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "UsageBatchIn",
        "type": "object"
      },
      "UsageCheckOut": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "soft_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soft Limit"
          },
          "soft_limit_breached": {
            "title": "Soft Limit Breached",
            "type": "boolean"
          },
          "used": {
            "title": "Used",
            "type": "integer"
          },
          "within_limit": {
            "title": "Within Limit",
            "type": "boolean"
          }
        },
        "required": [
          "metric",
          "used",
          "limit",
          "remaining",
          "within_limit",
          "soft_limit",
          "soft_limit_breached"
        ],
        "title": "UsageCheckOut",
        "type": "object"
      },
      "UsageEventIn": {
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "properties": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties"
          },
          "quantity": {
            "default": 1,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "metric"
        ],
        "title": "UsageEventIn",
        "type": "object"
      },
      "UsageResultOut": {
        "properties": {
          "deduplicated": {
            "default": false,
            "title": "Deduplicated",
            "type": "boolean"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "within_limit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Within Limit"
          }
        },
        "required": [
          "metric",
          "quantity",
          "total"
        ],
        "title": "UsageResultOut",
        "type": "object"
      },
      "UsageSummaryOut": {
        "properties": {
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/UsageCheckOut"
            },
            "title": "Metrics",
            "type": "array"
          },
          "period": {
            "title": "Period",
            "type": "string"
          }
        },
        "required": [
          "period",
          "metrics"
        ],
        "title": "UsageSummaryOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multi-tenant SaaS framework: tenancy, plans, entitlements, usage, billing.",
    "title": "Synapse SaaS Framework",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/plans": {
      "get": {
        "operationId": "list_plans_v1_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanRead"
                  },
                  "title": "Response List Plans V1 Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Plans",
        "tags": [
          "subscriptions"
        ]
      }
    },
...
    "/v1/subscription/cancel": {
      "post": {
        "operationId": "cancel_subscription_v1_subscription_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Subscription",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/v1/subscription/change": {
      "post": {
        "operationId": "change_plan_v1_subscription_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Plan",
        "tags": [
          "subscriptions"
        ]
      }
    },
...
    "/v1/subscription/trial": {
      "post": {
        "operationId": "start_trial_v1_subscription_trial_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrialStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Trial",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/v1/usage/check": {
      "get": {
        "operationId": "check_usage_v1_usage_check_get",
        "parameters": [
          {
            "in": "query",
            "name": "metric",
            "required": true,
            "schema": {
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "quantity",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Quantity",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageCheckOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Usage",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/consume": {
      "post": {
        "description": "Meter + enforce ONE event. 402 with upgrade hints on breach.\n\nMore than one event is a 422 (it used to silently drop all but the first);\nuse `/usage/consume-batch` for all-or-nothing batches.",
        "operationId": "consume_v1_usage_consume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageBatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consume",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/consume-batch": {
      "post": {
        "description": "Meter + enforce a batch atomically: the first breach 402s and NOTHING in\nthe batch is counted (the request transaction rolls back).",
        "operationId": "consume_batch_v1_usage_consume_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageBatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsageResultOut"
                  },
                  "title": "Response Consume Batch V1 Usage Consume Batch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consume Batch",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/events": {
      "post": {
        "description": "Meter usage. Recording never blocks (soft path).",
        "operationId": "record_events_v1_usage_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageBatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsageResultOut"
                  },
                  "title": "Response Record Events V1 Usage Events Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Events",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/gauge": {
      "post": {
        "description": "Set (`value`) or move (`delta`) a gauge metric \u2014 seats, projects, bytes.\n\nGauges are levels, not flows: they never reset with the billing period and\nare never metered through /usage/events or /usage/consume. A positive\n`delta` is capacity-checked: **402** with upgrade hints when it would\nexceed the plan's cap, and the level is left untouched.",
        "operationId": "set_gauge_v1_usage_gauge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GaugeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Gauge",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/summary": {
      "get": {
        "operationId": "usage_summary_v1_usage_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage Summary",
        "tags": [
          "usage"
        ]
      }
    },
...
  }
}